+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Updates README to indicate that Andoid SDK <21 is no longer supported.
+* Updates minimum supported SDK version to Flutter 3.24/Dart 3.5.
+
 ## 2.9.5
 
 * Fixes layout issue caused by `Transform.rotate` not affecting space calculation.